EAMxx: share mapping data across horiz remappers
Namely, coarse/ov_coarse grids as well as the
CRS matrix data. This is not just a memory saving,
but most importantly a runtime saving.
