+++ components/eamxx/src/share/CMakeLists.txt
@@ -24,6 +24,7 @@ set(SHARE_SRC
   grid/remap/abstract_remapper.cpp
   grid/remap/coarsening_remapper.cpp
   grid/remap/horiz_interp_remapper_base.cpp
+  grid/remap/horiz_interp_remap_data.cpp
   grid/remap/refining_remapper_p2p.cpp
   grid/remap/vertical_remapper.cpp
   property_checks/property_check.cpp
