+++ components/eamxx/src/share/grid/remap/horiz_interp_remapper_base.hpp
@@ -2,6 +2,7 @@
 #define SCREAM_HORIZ_INTERP_REMAPPER_BASE_HPP
 
 #include "share/grid/remap/abstract_remapper.hpp"
+#include "share/grid/remap/horiz_interp_remap_data.hpp"
 
 namespace scream
 {
@@ -16,18 +17,12 @@ namespace scream
 
 class HorizInterpRemapperBase : public AbstractRemapper
 {
-protected:
-  enum class InterpType {
-    Refine,
-    Coarsen
-  };
-
 public:
   HorizInterpRemapperBase (const grid_ptr_type& fine_grid,
                            const std::string& map_file,
                            const InterpType type);
 
-  virtual ~HorizInterpRemapperBase () = default;
+  ~HorizInterpRemapperBase ();
 
   FieldLayout create_src_layout (const FieldLayout& tgt_layout) const override;
   FieldLayout create_tgt_layout (const FieldLayout& src_layout) const override;
@@ -69,31 +64,10 @@ protected:
     EKAT_ERROR_MSG ("HorizInterpRemapperBase only supports fwd remapping.\n");
   }
 
-  using gid_type = AbstractGrid::gid_type;
   using KT = KokkosTypes<DefaultDevice>;
 
   template<typename T>
   using view_1d = typename KT::template view_1d<T>;
-
-  struct Triplet {
-    // Note: unfortunately, C++17 does not support emplace-ing POD
-    //       types as aggregates unless a ctor is declared. C++20 does though.
-    Triplet () = default;
-    Triplet(const gid_type rr, const gid_type cc, const Real ww)
-      : row(rr), col(cc), w(ww) {}
-    gid_type row;
-    gid_type col;
-    Real  w;
-  };
-
-  std::vector<Triplet>
-  get_my_triplets (const std::string& map_file) const;
-
-  void create_coarse_grids (const std::vector<Triplet>& triplets);
-
-  // Not a const ref, since we'll sort the triplets according to
-  // how row gids appear in the coarse grid
-  void create_crs_matrix_structures (std::vector<Triplet>& triplets);
 
   void create_ov_fields ();
 
@@ -131,6 +105,10 @@ public:
   view_1d<int>    m_row_offsets;
   view_1d<int>    m_col_lids;
   view_1d<Real>   m_weights;
+
+  // Keep track of this, since we need to tell the remap data repo
+  // we are releasing the data for our map file.
+  std::string     m_map_file;
 
   InterpType      m_type;
